+++ core/src/test/java/com/example/demo/MovieRecommendationControllerImpTest.java
@@ -24,7 +24,7 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class MovieRecommendationControllerTest {
+public class MovieRecommendationControllerImpTest {
     @Autowired
     private MovieRecommendationService movieRecommendationService;
     @Autowired
